move user data export from server action to API route
