+++ docs/auth.md
@@ -28,10 +28,10 @@
 ## Data Export
 - Users can export all their personal data from `/settings/account` (GDPR Art. 20 — right to data portability)
 - Exported as a JSON file named `kosmo-data-export-YYYY-MM-DD.json`
-- Rate-limited to once per hour via `lastDataExportAt` on `User`
+- Rate-limited to once per hour via `lastDataExportAt` on `User`; cooldown constant lives in `lib/constants.ts`
 - The cooldown check is atomic: a raw SQL `UPDATE ... WHERE` sets `lastDataExportAt` only if the cooldown has passed, preventing concurrent requests from bypassing the limit
 - If the export query fails, `lastDataExportAt` is reset to `null` so the user isn't locked out
-- Server action: `exportUserData()` in `lib/actions/settings.ts`
+- API route: `GET /api/export-data` (`app/api/export-data/route.ts`) — returns the JSON file directly with `Content-Disposition: attachment`, avoiding server-action response-size limits for large accounts
 - Export includes: profile, preferences, posts (with communities & tags), comments, votes, drafts, community memberships, feed memberships, follows, feedbacks, feedback votes, owned communities, owned feeds, sessions (IP address, user agent)
 
 ## Models
