+++ lib/constants.ts
@@ -1,0 +1,1 @@
+export const EXPORT_COOLDOWN_MS = 60 * 60 * 1000 // 1 hour
